fix integer overflow

+++ gui/src/main/groovy/com/muwire/gui/DownloadProgressRenderer.groovy
@@ -23,7 +23,7 @@ class DownloadProgressRenderer extends DefaultTableCellRenderer {
         int percent = -1
         if (pieces != 0)
             percent = (int)(done * 100.0 / pieces)
-        long size = d.pieceSize * pieces
+        long size = d.pieceSize * ((long) pieces)
         String totalSize = DataHelper.formatSize2Decimal(size, false) + "B"
         setText(String.format("%2d", percent) + "% of ${totalSize} ($done/$pieces pcs)".toString())
